# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "kids-r-us",
        "description": "E-commerce web api",
        "contact": {
            "name": "lleraabi@gmail.com",
            "email": "lleraabi@gmail.com"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/products/photo/{product}": {
            "post": {
                "summary": "Guarda una nueva foto de un producto",
                "description": "Guarda una nueva foto de un producto",
                "operationId": "f308b89f32d3422c32f1a14ec4abac53",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Id del producto",
                        "required": true,
                        "allowEmptyValue": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "photo": {
                                        "description": "Foto a subir",
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/products/photo/{photo}": {
            "delete": {
                "summary": "Elimina una foto de un producto",
                "description": "Elimina una foto de un producto",
                "operationId": "527172519995070865a72289739a27ea",
                "parameters": [
                    {
                        "name": "photo",
                        "in": "path",
                        "description": "Id de la foto a eliminar",
                        "required": true,
                        "allowEmptyValue": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
        "/api/user/profile-information": {
            "put": {
                "summary": "Endpoint para actualizar los datos del usuario autenticado.",
                "description": "Endpoint para actualizar los datos del usuario autenticado.",
                "operationId": "66ac266506c04157f941941513599add",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nombre de usuario",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Dirección de correo",
                                        "type": "string"
                                    },
                                    "photo": {
                                        "description": "Foto de perfil",
                                        "nullable": true
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Juan",
                                    "email": "juan@gmail.com",
                                    "photo": "Alguna foto de perfil menor de 2048 bytes..."
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
...
        "/api/reviews/{product}": {
            "get": {
                "summary": "Obtiene las reseñas de un producto",
                "description": "Obtiene las reseñas de un producto.",
                "operationId": "663189e0421b3556e88a12c4308f4c4c",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Id del producto",
                        "required": true,
                        "allowEmptyValue": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
        "/api/reviews": {
            "post": {
                "summary": "Guarda una nueva reseña",
                "description": "Guarda una nueva reseña de un producto.",
                "operationId": "154c71f5f72b459d998f5ea451bd4c96",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/Review"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/Review"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reviews/{review}": {
            "put": {
                "summary": "Actualiza una reseña",
                "description": "Actualiza una reseña del usaurio autenticado.",
                "operationId": "ac24d04d9d6d1e7753ce24c811a3cbb0",
                "parameters": [
                    {
                        "name": "review",
                        "in": "path",
                        "description": "Id de la reseña",
                        "required": true,
                        "allowEmptyValue": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/Review"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            },
            "delete": {
                "summary": "Elimina un reseña",
                "description": "elimina una reseña del usuario autenticado",
                "operationId": "7d89060a6133216c7215b8624ce930fe",
                "parameters": [
                    {
                        "name": "review",
                        "in": "path",
                        "description": "Id de la reseña",
                        "required": true,
                        "allowEmptyValue": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
        "/api/products/search": {
            "get": {
                "summary": "Busca los productos por las características",
                "description": "Realiza una búsqueda de los productos por las características dadas.",
                "operationId": "8071976f63ed39270388f29890d79ac1",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre del producto a buscar"
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Categoría del producto"
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Etiquetas que pueda tener el producto"
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Precio"
                    },
                    {
                        "name": "stock",
                        "in": "query",
                        "description": "Existencias"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ProductPaginatorSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/search-count": {
            "get": {
                "summary": "Busca los productos por las características",
                "description": "Realiza una búsqueda de los productos por las características dadas.",
                "operationId": "6e324ec39e03bccc57ce1d4792e8c848",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre del producto a buscar"
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Categoría del producto"
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Etiquetas que pueda tener el producto"
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Precio"
                    },
                    {
                        "name": "stock",
                        "in": "query",
                        "description": "Existencias"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna la cantidad de productos encontrados"
                    }
                }
            }
        },
...
    },
...
}